+++ src/components/Note/CreateInput/LocationInput/index.module.scss
