+++ src/components/Note/CreateInput/PictureInput/index.module.scss
