+++ src/components/Note/common/BottomSheet/index.module.scss
@@ -48,8 +48,8 @@
   height: 370px;
 }
 
-.bottomSheetFooter {
-}
+// .bottomSheetFooter {
+// }
 
 .bottomSheetSaveButton {
   width: 100%;
